requirements.txt
python-dotenv
requests
feedparser
transformers
torch

